Initial setup
Setup an initial environment so that a basic JIT can be hooked
into the Lua VM.

This requires making a few minor changes to the Lua VM as well as
creating some new Makefiles that will operate the build process.

For testing purposes a "stub" JIT is implemented that never compiles
anything.