Implement ADD and partially implement SETTABUP
Create an initial implementation of `ADD` (only supports integers).
This implementation is generic enough to only require small changes
to support other arithmetic operators.

Partially implement the `SETTABUP`. Due to the complexity of the
macros that have to be unwrapped, creating a full implementation will
take time. This is an initial attempt at getting parts of the structure
implemented (though commented out).