Implement proper TValue tagged union support
To properly overlay the `TValue` tagged union, we define a different
struct for each possible type stored in the tagged union. There is one
struct for each union field. The value field of each struct corresponds
to one of the fields of the union. The tag field has the same type in
all structs.