Resolve a runtime compilation assertion
When generating code for a function I was getting a vmState != NULL
assertion.  The issue was with a JMP instruction directly following a
RETURN. The JMP was unreachable code that just did a jump to the
unreachable RETURN at the end of the function.

The solution easiest solution is to switch to the main loop which
generates code for the instructions to use the MethodBuilder work list.
Using the work list ensures code is only generated for each instruction
once and that you do not visit instructions before they should be.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>