Fix load_nil
Load_nil is supposed to load b number of nils to the registers starting
at ra.  Currently the code was load nil to ra n times.  The fix is to
update ra everytime through the loop.  I modified the test as well to
check that the right registers were set to nil.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>