Issue #14 handle OP_VARARG
Implement IL for most of the operations required for OP_VARARG.
Currently the code is calling a helper to check the stack / perform any
GC work.  A further change will be to do the stack check in IL and call
the GC helper if it is required.  I added tests for this opcode but I
was not able to generate code to test the case where b is > 0 so the
extra slots get nil values. 

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>