Stack growth not handled on OP_CALL
When compiled code calls a function this called function may cause the
lua stack to grow or move. If the stack is modified the compiled code
needs to reload "base" from the CallInfo.

The issue can be viewed by running fib(40) as it will cause a runtime
error due to reading the wrong values.  With this fix fib(40) runs to
completion.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>