Do not store "ra" for every opcode
A lot of opcodes do not use "ra" and in some cases create another
copy of "ra" locally.  This causes a lot of extra IL to be
generated.  Instead only create an IlValue for "ra" in the handler
for the opcode if it uses "ra".  Removing this cleans up the IL
and makes it easier to read verbose JIT logs.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>